requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
pydantic>=2.0
numpy>=1.24
torch>=2.0
torch-geometric>=2.5
rdkit-pypi>=2022.9
scipy>=1.11
requests>=2.31
httpx>=0.27
